==> DOW分鐘漲跌預測_保留關鍵檔案/08.IB_RAW資料進行合併.ipynb <==
import pandas as pd
import numpy as np

商品 = "ZN"

IB09_df = pd.read_parquet(f"datas/{商品}/IB_{商品}_5min-20240701-20240919.pqt")
IB12_df = pd.read_parquet(f"datas/{商品}/IB_{商品}_5min-20240923-20241030.pqt")
IB09_df["datetime"] = pd.to_datetime(IB09_df.index)
IB12_df["datetime"] = pd.to_datetime(IB12_df.index)

# 合併IB
# 移除IB09_df['Datetime']不在15:00:00-04:00:00的row
IB09_df = IB09_df[
    (IB09_df["datetime"].dt.time >= pd.to_datetime("15:00:00").time())
    | (IB09_df["datetime"].dt.time <= pd.to_datetime("04:00:00").time())
]
# 移除IB12_df['Datetime']不在15:00:00-04:00:00的row
IB12_df = IB12_df[
    (IB12_df["datetime"].dt.time >= pd.to_datetime("15:00:00").time())
    | (IB12_df["datetime"].dt.time <= pd.to_datetime("04:00:00").time())
]
# 合併IB09_df和IB12_df
IB_df = pd.concat([IB09_df, IB12_df])

# 移除IB_df['datetime']重複的row
IB_df = IB_df.drop_duplicates(subset=["datetime"])

關於直接整理ES 5min資料集

RTY_df = IB_df.copy()
# 假設你已有 DataFrame 'df'，並包含 'close' 欄位以及以5分鐘為間隔的索引
# ES_df["ES 5分鐘震幅"] = (ES_df["high"] - ES_df["low"]) * 10000 / ES_df["low"]
# 新增5分鐘、10分鐘、30分鐘、60分鐘的漲跌幅
# ES_df["ES 5分鐘漲跌幅"] = (ES_df["close"] - ES_df["open"]) * 10000 / ES_df["close"]
# ES_df["ES 10分鐘漲跌幅"] = ES_df["close"].pct_change(periods=2) * 10000  # 10分鐘

RTY_df[f"{商品} 30分鐘漲跌幅"] = RTY_df["close"].pct_change(periods=6) * 10000  # 30分鐘

RTY_df[f"{商品} 60分鐘漲跌幅"] = (
    RTY_df["close"].pct_change(periods=12) * 10000
)  # 60分鐘

# 新增30分鐘和60分鐘的平均收盤價距離現在的漲跌幅
# ES_df["ES 30分鐘平均的close比"] = (ES_df["close"]-ES_df["close"].rolling(window=6).mean() )*10000 / ES_df["close"]

# ES_df["ES 60分鐘平均的close比"] = (
#     (ES_df["close"] - ES_df["close"].rolling(window=12).mean()) * 10000 / ES_df["close"]
# )
# 處理成交量
# ES_df["ES 60分鐘平均的volume"] = (
#     (ES_df["volume"] - ES_df["volume"].rolling(window=6).mean())
# )
# ES_df["ES 60分鐘平均的count"] = (
#     (ES_df["count"] - ES_df["count"].rolling(window=6).mean())
# )
# ES_df["ES 每筆成交量"] = ES_df["volume"] / ES_df["count"]

RTY_df=RTY_df.drop(["high", "low", "open","datetime"], axis=1)


RTY_df = RTY_df.rename(
    {

        "open": f"{商品} open",
        "close": f"{商品} close",
        "volume": f"{商品} volume",
        "count": f"{商品} count",
    },
    axis="columns",
)

RTY_df

RTY_df.to_csv(f"datas/{商品}/{商品}_0701-1030_5min含筆數與張數.csv")

關於將YM 1min資料集合並成5min資料集

# 將數據重新取樣為每5分鐘
def resample_5min(df):
    # 計算每5分鐘的成交總張數、成交總筆數
    df_resampled = df.resample("5T").agg(
        {"volume": ["sum", "mean", "max", "min"], "count": ["sum", "std"]}
    )
    df_resampled.columns = [
        "YM五分成交總量",
        "YM五分內平均1min成交量",
        "YM五分內最大1min成交量",
        "YM五分內最小1min成交量",
        "YM五分內成交總筆數",
        "YM五分內成交筆數標準差",
    ]

    # 計算每筆成交的平均張數
    df_resampled["YM五分內平均每筆張數"] = (
        df_resampled["YM五分成交總量"] / df_resampled["YM五分內成交總筆數"]
    )

    # 計算大於120張的成交次數與大於80筆的成交次數
    df_resampled["YM五分內每min大於120張的次數"] = (
        df["volume"].resample("5T").apply(lambda x: (x > 120).sum())
    )
    df_resampled["YM五分內每min大於80筆的次數"] = (
        df["count"].resample("5T").apply(lambda x: (x > 80).sum())
    )

    # 處理可能的 NaN 值
    df_resampled.fillna(0, inplace=True)

    return df_resampled


df_5min = resample_5min(IB_df)
df_5min

df_5min.describe()

df_5min.to_csv("datas/YM/YM_0701-1030_5min含筆數與張數.csv")
